Some structural fixes + common files + styles for guest + polish
