+++ package.json
@@ -33,7 +33,6 @@
     "null-loader": "^3.0.0",
     "postcss-loader": "^3.0.0",
     "sass-loader": "^8.0.0",
-    "startserver": "^1.3.8",
     "style-loader": "^1.0.0",
     "url-loader": "^2.2.0",
     "webpack": "^4.41.2",
