+++ src/html/guest.html
@@ -13,7 +13,7 @@
 
 <header id="header">
     <div>
-        <h1>Guest panel</h1>
+        <h1>Panel gościa</h1>
     </div>
     <div>
         <div id="header-user">
@@ -23,9 +23,9 @@
             </div>
 
             <div id="header-userActions">
-                <div class="header-userActions-action" title="Logout">
-                    <img src="../icons/logout.png" alt="Logout" />
-                    <span>Logout</span>
+                <div class="header-userActions-action" title="Wyloguj">
+                    <img src="../icons/logout.png" alt="Wyloguj" />
+                    <span>Wyloguj</span>
                 </div>
             </div>
         </div>
@@ -52,10 +52,10 @@
 </main>
 
 <section class="section" id="fillForm">
-    <div class="sectionBack">
+    <div id="fillForm-back" class="sectionBack">
         <div>
             <div class="sectionBack-content">
-                <img src="../icons/arrow-left.png" alt="Back" />
+                <img src="../icons/arrow-left.png" alt="Powrót" />
                 <span>Powrót</span>
             </div>
         </div>
@@ -69,9 +69,9 @@
         <div id="fillForm-list">
             <div id="fillForm-list-table">
                 <div name="title">
-                    <div>No.</div>
-                    <div>Name</div>
-                    <div>Questions</div>
+                    <div>Nr.</div>
+                    <div>Nazwa</div>
+                    <div>Pytania</div>
                     <div></div>
                 </div>
             </div>
