+++ src/scss/guest.fillForm.scss
@@ -1,4 +1,4 @@
-#showForms-content {
+#fillForm-content {
     width: 100%;
     max-width: 1000px;
     height:600px;
@@ -7,10 +7,11 @@
     box-shadow: 0 0 5px 1px #ccc;
     overflow: auto;
 }
-    #showForms-list {
+    #fillForm-list {
         width:100%;
+        height:100%;
     }
-        #showForms-list-table {
+        #fillForm-list-table {
             width: 100%;
 
             > div { // row
@@ -44,7 +45,7 @@
                         text-align: center;
                     }
                     &:nth-of-type(2) {
-                        width: calc(100% - 10px - 50px - 100px - 40px - 40px - 40px);
+                        width: calc(100% - 10px - 50px - 100px - 50px);
                         white-space: nowrap;
                         padding: 0 5px;
                         text-overflow: ellipsis;
@@ -56,12 +57,7 @@
                     }
                     &:nth-of-type(4) {
                         width: 40px;
-                    }
-                    &:nth-of-type(5) {
-                        width: 40px;
-                    }
-                    &:last-of-type {
-                        width: 40px;
+                        padding-right: 10px;
                     }
 
                     > img {
@@ -80,8 +76,9 @@
             }
         }
 
-    #showForms-form {
+    #fillForm-fill {
         display: none;
+        height:100%;
 
         > h3 {
             background: $basecolor;
@@ -94,26 +91,27 @@
             color: #fff;
         }
     }
-        #showForms-form-content {
+        #fillForm-form-content {
             width:calc(100% - 40px);
+            height: calc(100% - 73px - 50px - 20px);
             padding: 10px 20px;
+            overflow: auto;
         }
 
+        #fillForm-form-buttons {
+            width:calc(100% - 40px);
+            height:50px;
+            padding: 0 20px;
+            background: #eee;
+            display: flex;
+            align-items: center;
+            justify-content: flex-end;
 
-    #showForms-form-buttons {
-        width:calc(100% - 40px);
-        height:50px;
-        padding: 0 20px;
-        background: #eee;
-        display: flex;
-        align-items: center;
-        justify-content: flex-start;
-
-        > button {
-            padding: 10px 20px;
-            background: $basecolor;
-            color: #fff;
-            border-radius: 5px;
-            border:none;
+            > button {
+                padding: 10px 20px;
+                background: $basecolor;
+                color: #fff;
+                border-radius: 5px;
+                border:none;
+            }
         }
-    }
