+++ src/scss/guest.scss
@@ -4,215 +4,7 @@ $basecolor: #39557C;
 @import 'common.header';
 
 @import 'common.section';
+@import 'common.form';
 
 @import 'guest.panel';
-
-
-    #fillForm-content {
-        width: 100%;
-        max-width: 1000px;
-        height:600px;
-        margin:0 auto;
-        background: #fff;
-        box-shadow: 0 0 5px 1px #ccc;
-        overflow: auto;
-    }
-        #fillForm-list {
-            width:100%;
-            height:100%;
-        }
-            #fillForm-list-table {
-                width: 100%;
-
-                > div { // row
-                    width:100%;
-                    height: 40px;
-                    font-family: 'Open Sans', sans-serif;
-                    line-height: 40px;
-                    overflow: hidden;
-
-                    &[name="title"] > div {
-                        background: #39557C;
-                        color: #fff;
-                        font-size: 1em;
-                    }
-
-                    &:nth-of-type(even) {
-                        background: #eee;
-                    }
-
-                    > div { // cell
-                        height: 40px;
-                        font-size: 1em/40px;
-                        float: left;
-
-                        &:first-of-type {
-                            width: 50px;
-                            text-align: center;
-                        }
-                        &:nth-of-type(2) {
-                            width: calc(100% - 10px - 50px - 100px - 40px);
-                            white-space: nowrap;
-                            padding: 0 5px;
-                            text-overflow: ellipsis;
-                            overflow: hidden;
-                        }
-                        &:nth-of-type(3) {
-                            width: 100px;
-                            text-align: center;
-                        }
-                        &:nth-of-type(4) {
-                            width: 40px;
-                        }
-
-                        > img {
-                            width: 24px;
-                            height:24px;
-                            margin: 4px;
-                            padding: 4px;
-                            cursor: pointer;
-                            border-radius: 50%;
-
-                            &:hover {
-                                background: #ccc;
-                            }
-                        }
-                    }
-                }
-            }
-
-        #fillForm-fill {
-            display: none;
-            height:100%;
-
-            > h3 {
-                background: $basecolor;
-                height:43px;
-                padding: 10px 0;
-                font-size: 1.8em;
-                font-family: 'Open Sans', sans-serif;
-                margin-bottom: 10px;
-                text-indent: 20px;
-                color: #fff;
-            }
-        }
-            #fillForm-form-content {
-                width:calc(100% - 40px);
-                height: calc(100% - 73px - 50px - 20px);
-                padding: 10px 20px;
-                overflow: auto;
-            }
-                .fillForm-form-oQuestion {
-                    width:100%;
-
-                    > p {
-                        font-size: 1.2em;
-                        font-family: 'Open Sans', sans-serif;
-                        font-weight: 700;
-                        margin-bottom: 10px;
-                    }
-
-                    > input {
-                        width: calc(100% - 20px);
-                        padding: 5px 10px;
-                        font-size: 1em;
-                        font-family: 'Open Sans', sans-serif;
-                        border: none;
-                        border-bottom: solid 2px #ccc;
-
-                        &:focus {
-                            border-bottom-color: #39557C;
-                        }
-                    }
-                }
-
-                .fillForm-form-question  {
-                    width:100%;
-                    margin-bottom: 20px;
-
-                    > p {
-                        font-size: 1.2em;
-                        font-family: 'Open Sans', sans-serif;
-                        font-weight: 700;
-                        margin-bottom: 10px;
-                    }
-                }
-
-                .fillForm-form-openQuestion,
-                .fillForm-form-numericalQuestion {
-                    > input {
-                        width: calc(100% - 20px);
-                        padding: 5px 10px;
-                        font-size: 1em;
-                        font-family: 'Open Sans', sans-serif;
-                        border: none;
-                        border-bottom: solid 2px #ccc;
-
-                        &:focus {
-                            border-bottom-color: #39557C;
-                        }
-                    }
-                }
-                .fillForm-form-closedQuestion {
-                    > label {
-                        width:100%;
-                        height:32px;
-                        display: block;
-                        cursor: pointer;
-
-                        > input {
-                            display: none;
-                        }
-
-                        > div {
-                            width:16px;
-                            height:16px;
-                            border:solid 2px #000;
-                            margin: 6px 0;
-                            border-radius: 50%;
-                            display: block;
-                            float: left;
-
-                            > div {
-                                width:10px;
-                                height:10px;
-                                margin:3px;
-                                background: #000;
-                                border-radius: 50%;
-                                transform: scale(0);
-                                transition: transform 0.3s;
-                            }
-                        }
-
-                        > input:checked ~ div > div {
-                            transform: scale(1);
-                        }
-
-                        > span {
-                            width:calc(100% - 20px - 20px);
-                            font-size: 1em;
-                            line-height: 32px;
-                            margin-left: 20px;
-                            font-family: 'Open Sans', sans-serif;
-                            display: block;
-                            float: left;
-                        }
-                    }
-                }
-            #fillForm-form-buttons {
-                width:calc(100% - 40px);
-                height:50px;
-                padding: 0 20px;
-                background: #eee;
-                display: flex;
-                align-items: center;
-                justify-content: flex-end;
-
-                > button {
-                    padding: 10px 20px;
-                    background: $basecolor;
-                    color: #fff;
-                    border-radius: 5px;
-                    border:none;
-                }
-            }
+@import 'guest.fillForm';
